+++ src/libged/track.c
@@ -197,7 +197,7 @@ ged_track(struct ged *gedp, int argc, const char *argv[])
     /* base name */
     arg = 1;
     grpname = bu_strdup(argv[arg]);
-    grpname_len = strlen(grpname);
+    grpname_len = (int)strlen(grpname);
     len = grpname_len + 1 + extraChars;
     solname = bu_malloc(len, "solid name");
     regname = bu_malloc(len, "region name");
